move 'UpcomingAppointmentList' and unit test to 'pages' directory
