+++ src/applications/vaos/appointment-list/pages/UpcomingAppointmentsPage/UpcomingAppointmentsPage.jsx
@@ -1,24 +1,24 @@
+import { recordEvent } from '@department-of-veterans-affairs/platform-monitoring/exports';
+import classNames from 'classnames';
+import moment from 'moment';
 import React, { useEffect } from 'react';
 import { shallowEqual, useDispatch, useSelector } from 'react-redux';
-import { recordEvent } from '@department-of-veterans-affairs/platform-monitoring/exports';
-import moment from 'moment';
-import classNames from 'classnames';
 import { useHistory } from 'react-router-dom';
-import InfoAlert from '../../components/InfoAlert';
-import { getUpcomingAppointmentListInfo } from '../redux/selectors';
-import { FETCH_STATUS, GA_PREFIX } from '../../utils/constants';
-import { groupAppointmentByDay } from '../../services/appointment';
-import NoAppointments from './NoAppointments';
-import { scrollAndFocus } from '../../utils/scrollAndFocus';
+import InfoAlert from '../../../components/InfoAlert';
+import { selectFeatureBreadcrumbUrlUpdate } from '../../../redux/selectors';
+import { groupAppointmentByDay } from '../../../services/appointment';
+import { FETCH_STATUS, GA_PREFIX } from '../../../utils/constants';
+import { scrollAndFocus } from '../../../utils/scrollAndFocus';
+import UpcomingAppointmentLayout from '../../components/AppointmentsPage/UpcomingAppointmentLayout';
+import BackendAppointmentServiceAlert from '../../components/BackendAppointmentServiceAlert';
+import NoAppointments from '../../components/NoAppointments';
 import {
   fetchFutureAppointments,
   startNewAppointmentFlow,
-} from '../redux/actions';
-import { selectFeatureBreadcrumbUrlUpdate } from '../../redux/selectors';
-import UpcomingAppointmentLayout from './AppointmentsPage/UpcomingAppointmentLayout';
-import BackendAppointmentServiceAlert from './BackendAppointmentServiceAlert';
+} from '../../redux/actions';
+import { getUpcomingAppointmentListInfo } from '../../redux/selectors';
 
-export default function UpcomingAppointmentsList() {
+export default function UpcomingAppointmentsPage() {
   const history = useHistory();
   const dispatch = useDispatch();
   const {
